refactor: Change deposit points and weight data type to float
Updates the data type for deposit points and weight from integer to float across the navigation graph and related screens.

- In `AuthenticatedNavHost`, the `navArgument` types for the `SuccessDeposit` route are changed from `NavType.IntType` to `NavType.FloatType`.
- In `SuccessDepositScreen`, the `points` and `weight` parameters are updated from `Int` to `Float` to correctly handle the floating-point values passed through navigation.

+++ app/src/main/java/com/ahargunyllib/growth/presentation/view/authenticated/SuccessDepositScreen.kt
@@ -30,8 +30,8 @@ import com.ahargunyllib.growth.presentation.ui.navigation.nav_obj.AuthenticatedN
 @Composable
 fun SuccessDepositScreen(
     authenticatedNavController: NavController,
-    points: Int,
-    weight: Int
+    points: Float,
+    weight: Float
 ) {
     Surface(
         modifier = Modifier.fillMaxSize(),
@@ -196,7 +196,7 @@ private fun InfoItem(icon: @Composable () -> Unit, text: String) {
 fun SuccessDepositScreenPreview() {
     SuccessDepositScreen(
         authenticatedNavController = rememberNavController(),
-        points = 20,
-        weight = 2
+        points = 20F,
+        weight = 2F
     )
 }
